Fetch command working, and refactor fetch/pull
