+++ src/lib/src/core/v0_10_0/entries.rs
@@ -19,7 +19,6 @@ use os_path::OsPath;
 
 use crate::core::df;
 use crate::core::v0_10_0::cache::cachers;
-use crate::core::v0_10_0::index;
 use crate::core::v0_10_0::index::ObjectDBReader;
 use crate::core::v0_10_0::index::{CommitDirEntryReader, CommitEntryReader, CommitReader};
 use crate::model::{
