+++ src/lib/src/repositories/pull.rs
@@ -51,27 +51,6 @@ pub async fn pull_remote_branch(
     }
 }
 
-/// Pull a remote repository, defaults to origin/main
-pub async fn pull_remote_repo(
-    repo: &LocalRepository,
-    remote_repo: &RemoteRepository,
-    opts: &PullOpts,
-) -> Result<(), OxenError> {
-    // If no remote branch is specified, use the default
-    let rb = RemoteBranch {
-        remote: DEFAULT_REMOTE_NAME.to_string(),
-        branch: DEFAULT_BRANCH_NAME.to_string(),
-    };
-    match repo.min_version() {
-        MinOxenVersion::V0_10_0 => {
-            core::v0_10_0::pull::pull_remote_repo(repo, remote_repo, &rb, opts).await
-        }
-        MinOxenVersion::V0_19_0 => {
-            core::v0_19_0::pull::pull_remote_repo(repo, remote_repo, &rb, opts).await
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use std::path::Path;
@@ -597,7 +576,7 @@ mod tests {
                 assert_eq!(cloned_num_files, 5);
 
                 // Switch to main branch and pull
-                repositories::fetch(&cloned_repo).await?;
+                repositories::fetch(&cloned_repo, false).await?;
                 repositories::checkout(&cloned_repo, "main").await?;
 
                 let cloned_num_files = util::fs::rcount_files_in_dir(&cloned_repo.path);
@@ -675,7 +654,7 @@ mod tests {
                 assert_eq!(cloned_num_files, 2);
 
                 // Switch to main branch and pull
-                repositories::fetch(&cloned_repo).await?;
+                repositories::fetch(&cloned_repo, false).await?;
 
                 repositories::checkout(&cloned_repo, branch_name).await?;
